# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Certbot API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/acme/accounts/{account_id}/certificates": {
            "get": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "List available certificates in an ACME account",
                "description": "",
                "operationId": "listCertificates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AcmeCertificate"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Create a new certificate in an ACME account",
                "description": "",
                "operationId": "createCertificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name of new certificate",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subjects",
                        "in": "formData",
                        "description": "list of subjects for this certificate, first is CN, following are subject alternative names",
                        "required": true,
                        "type": "array",
                        "items": {
                            "description": "sibject cn or san ex: sub.domain.com",
                            "type": "string"
                        }
                    },
                    {
                        "name": "request",
                        "in": "formData",
                        "description": "optional externally generated PKCS10 certificate signing request -----BEGIN CERTIFICATE REQUEST-----\\nBASE64\\n-----END CERTIFICATE REQUEST-----",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AcmeCertificate"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/acme/accounts/{account_id}/certificates/{certificate_id}": {
            "get": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Find certificate in accme account by ID",
                "description": "",
                "operationId": "getCertificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate in this account",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AcmeCertificate"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Update certificate in account by ID",
                "description": "",
                "operationId": "updateCertificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name of certificate",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "subjects",
                        "in": "formData",
                        "description": "list of subjects for this certificate, first is CN, following are subject alternative names",
                        "required": false,
                        "type": "array",
                        "items": {
                            "description": "sibject cn or san ex: sub.domain.com",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AcmeCertificate"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Delete certificate in account by id",
                "description": "",
                "operationId": "deleteCertificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/acme/accounts/{account_id}/certificates/{certificate_id}/generatekeys": {
            "post": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Generate new key pair for certificate",
                "description": "Keypair generation is required if you plan to generate a certificate signing request inside certbot",
                "operationId": "certificateGenerateKeys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/acme/accounts/{account_id}/certificates/{certificate_id}/generaterequest": {
            "post": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Generate new certificate signing request",
                "description": "This is only necessary if you did not load an externally generated CSR into the tool",
                "operationId": "certificateGenerateRequest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/acme/accounts/{account_id}/certificates/{certificate_id}/sign": {
            "post": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Sign this certificates request",
                "description": "You must have signing permissions for the owning account AND a valid CSR for an ACME ca to sign must be provided or generated",
                "operationId": "certificateSign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/acme/accounts/{account_id}/certificates/{certificate_id}/renew": {
            "post": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Renew this certificate",
                "description": "Before expiration a certificate can be renewed without re-verification provided its request signature has not changed",
                "operationId": "certificateRenew",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/acme/accounts/{account_id}/certificates/{certificate_id}/pkcs12": {
            "get": {
                "tags": [
                    "Acme Certificates"
                ],
                "summary": "Download certificate, chain, and private key PKCS12 format",
                "description": "",
                "operationId": "certificateDownloadPKCS12",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/x-pkcs12"
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of account",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "certificate_id",
                        "in": "path",
                        "description": "ID of certificate",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "optional password to encrypt pkcs12 file contents",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "AcmeCertificate": {
            "required": [
                "name",
                "subjects"
            ],
            "properties": {
                "id": {
                    "description": "Unique identifier for the account id",
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "description": "Name of this certificate",
                    "type": "string"
                },
                "subjects": {
                    "description": "array of at least one CN followed by subject alternative names for this certificate",
                    "type": "array",
                    "items": []
                },
                "expires": {
                    "description": "Date the current certificate expires if applicable",
                    "type": "string",
                    "format": "date-format"
                },
                "status": {
                    "description": "status of this certificate, new unsigned signed etc",
                    "type": "string",
                    "enum": [
                        "new",
                        "unsigned",
                        "signed"
                    ]
                },
                "created_at": {
                    "description": "Date this interaction was created",
                    "type": "string",
                    "format": "date-format"
                },
                "updated_at": {
                    "description": "Date this interaction was last updated",
                    "type": "string",
                    "format": "date-format"
                },
                "deleted_at": {
                    "description": "Date this interaction was deleted",
                    "type": "string",
                    "format": "date-format"
                }
            }
        },
...
    }
}